data_pipeline/requirements.txt
requests
psycopg2-binary
pandas
sqlalchemy
pymssql